workspace/libs/rvt-adapter/pyproject.toml
[project]
name = "rvt-adapter"
version = "0.1.0"
description = "Shared helpers for loading RVT configs, checkpoints, and tensors"
requires-python = ">=3.11,<3.12"
dependencies = [
    "evio-core",
    "hydra-core>=1.3.2",
    "omegaconf>=2.3",
    "polars>=0.20.0",
    "torch>=2.0",
]

[tool.uv.sources]
evio-core = { workspace = true }

[build-system]
requires = ["hatchling>=1.25"]
build-backend = "hatchling.build"
